# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Valley MAS 创作者内容平台 API 文档\n支持创作者注册、口令管理、资源上传下载等功能",
    "title": "Valley MAS API",
    "contact": {
      "name": "API Support",
      "email": "support@valley-mas.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/admin/users": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "支持分页、关键词搜索、平台筛选、角色筛选",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 用户管理"],
        "summary": "获取用户列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "string",
            "description": "关键词搜索（昵称/OpenID）",
            "name": "keyword",
            "in": "query"
          },
          {
            "enum": ["wechat", "douyin", "all"],
            "type": "string",
            "description": "平台筛选",
            "name": "platform",
            "in": "query"
          },
          {
            "enum": ["user", "creator", "admin"],
            "type": "string",
            "description": "角色筛选",
            "name": "role",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "用户列表",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员创建新用户",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 用户管理"],
        "summary": "创建用户",
        "parameters": [
          {
            "description": "用户信息",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/admin/users/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "根据用户ID获取详细信息",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 用户管理"],
        "summary": "获取用户详情",
        "parameters": [
          {
            "type": "string",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "用户详情",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "用户不存在",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员更新用户信息",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 用户管理"],
        "summary": "更新用户信息",
        "parameters": [
          {
            "type": "string",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "用户信息",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "用户不存在",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员删除用户（软删除）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 用户管理"],
        "summary": "删除用户",
        "parameters": [
          {
            "type": "string",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "删除失败",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/admin/users/{id}/status": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "启用或禁用用户",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 用户管理"],
        "summary": "更新用户状态",
        "parameters": [
          {
            "type": "string",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "状态",
            "name": "status",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "isActive": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "用户不存在",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "model.User": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "douyinAvatar": {
          "description": "抖音头像",
          "type": "string"
        },
        "douyinCity": {
          "description": "抖音用户城市",
          "type": "string"
        },
        "douyinCountry": {
          "description": "抖音用户国家",
          "type": "string"
        },
        "douyinGender": {
          "description": "抖音性别: 0-未知 1-男 2-女",
          "type": "integer"
        },
        "douyinNickname": {
          "description": "抖音昵称",
          "type": "string"
        },
        "douyinOpenid": {
          "description": "抖音平台特有字段",
          "type": "string"
        },
        "douyinProvince": {
          "description": "抖音用户省份",
          "type": "string"
        },
        "douyinUnionid": {
          "description": "抖音开放平台账号唯一标识",
          "type": "string"
        },
        "email": {
          "description": "邮箱（可选）",
          "type": "string"
        },
        "id": {
          "description": "Snowflake ID (序列化为字符串)",
          "type": "integer"
        },
        "isActive": {
          "type": "boolean"
        },
        "nickname": {
          "type": "string"
        },
        "openid": {
          "description": "平台唯一标识",
          "type": "string"
        },
        "phone": {
          "description": "手机号（可选）",
          "type": "string"
        },
        "platform": {
          "description": "wechat, douyin, mini_app 等",
          "type": "string"
        },
        "role": {
          "description": "user, admin, creator",
          "type": "string"
        },
        "unionid": {
          "description": "同一主体下的唯一标识",
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "username": {
          "description": "管理后台登录字段",
          "type": "string"
        },
        "wechatOpenid": {
          "description": "微信平台特有字段（可选）",
          "type": "string"
        },
        "wechatUnionid": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT 认证，格式：Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}